libGwen: Replace comma separators with periods.
The reason behind this one is a doozy.  There is no spec that says you
cannot use commas for separators (and many MIME clients do this all the
time), but this breaks PHP's file upload handler because of the way
they check for the end of the boundary string -- see main/rfc1867.c:742
in PHP's source.  This fixes PHP file uploads by using periods instead
of commas for separators.